Refactor symbols page
